move code for decoding data into separate file
